Put GUASSIAN LOG parser & test into log.py file
