+++ iodata/fchk.py
@@ -19,135 +19,15 @@
 #
 # --
 # pragma pylint: disable=wrong-import-order,invalid-name
-"""Module for handling GAUSSIAN LOG and FCHK file formats."""
+"""Module for handling GAUSSIAN FCHK file format."""
 
 
 import numpy as np
 
 from typing import Dict, TextIO, Set, List
 
-from .utils import set_four_index_element
 
-
-__all__ = ['load_operators_g09', 'FCHKFile', 'load_fchk']
-
-
-def load_operators_g09(filename: str) -> Dict:
-    """Load several two- and four-index operators from a GAUSSIAN09 LOG file format.
-
-    Parameters
-    ----------
-    filename
-        The GAUSSIAN09 LOG filename.
-
-    Returns
-    -------
-    out : dict
-        Output dictionary may contain ``olp``, ``kin``, ``na`` and/or ``er`` keys and
-        corresponding values.
-
-    Notes
-    -----
-    1. The following two-index operators are loaded if present:
-       overlap, kinetic, nuclear attraction.
-    2. The following four-index operator is loaded if present: electrostatic repulsion.
-       In order to make all these matrices present in the GAUSSIAN LOG file, the following
-       commands must be used in the GAUSSIAN INPUT file:
-       ```scf(conventional) iop(3/33=5) extralinks=l316 iop(3/27=999)```
-
-    """
-    with open(filename) as f:
-        # First get the line with the number of orbital basis functions
-        for line in f:
-            if line.startswith('    NBasis ='):
-                nbasis = int(line[12:18])
-                break
-
-        # Then load the two- and four-index operators. This part is written such
-        # that it does not make any assumptions about the order in which these
-        # operators are printed.
-        result = {}
-        for line in f:
-            if line.startswith(' *** Overlap ***'):
-                result['olp'] = _load_twoindex_g09(f, nbasis)
-            elif line.startswith(' *** Kinetic Energy ***'):
-                result['kin'] = _load_twoindex_g09(f, nbasis)
-            elif line.startswith(' ***** Potential Energy *****'):
-                result['na'] = _load_twoindex_g09(f, nbasis)
-            elif line.startswith(' *** Dumping Two-Electron integrals ***'):
-                result['er'] = _load_fourindex_g09(f, nbasis)
-        return result
-
-
-def _load_twoindex_g09(f: TextIO, nbasis: int) -> np.ndarray:
-    """Load a two-index operator from a GAUSSIAN LOG file format.
-
-    Parameters
-    ----------
-    f
-        A GAUSSIAN LOG file object (in read mode).
-    nbasis
-        The number of atomic orbital basis functions.
-
-    Returns
-    -------
-    out: array_like
-        The output (nbasis, nbasis) array of operator.
-
-    """
-    result = np.zeros((nbasis, nbasis))
-    block_counter = 0
-    while block_counter < nbasis:
-        # skip the header line
-        next(f)
-        # determine the number of rows in this part
-        nrow = nbasis - block_counter
-        for i in range(nrow):
-            words = next(f).split()[1:]
-            for j in range(len(words)):
-                value = float(words[j].replace('D', 'E'))
-                result[i + block_counter, j + block_counter] = value
-                result[j + block_counter, i + block_counter] = value
-        block_counter += 5
-    return result
-
-
-def _load_fourindex_g09(f: TextIO, nbasis: int) -> np.ndarray:
-    """Load a four-index operator from a GAUSSIAN LOG file.
-
-    Parameters
-    ----------
-    f
-        A GAUSSIAN LOG file object (in read mode).
-    nbasis
-        The number of atomic orbital basis functions.
-
-    Returns
-    -------
-    out: array_like
-        The (nbasis, nbasis, nbasis, nbasis) array of operator.
-
-    """
-    result = np.zeros((nbasis, nbasis, nbasis, nbasis))
-    # Skip first six lines
-    for i in range(6):
-        next(f)
-    # Start reading elements until a line is encountered that does not start
-    # with ' I='
-    while True:
-        line = next(f)
-        if not line.startswith(' I='):
-            break
-        # print line[3:7], line[9:13], line[15:19], line[21:25], line[28:].replace('D', 'E')
-        i = int(line[3:7]) - 1
-        j = int(line[9:13]) - 1
-        k = int(line[15:19]) - 1
-        l = int(line[21:25]) - 1
-        value = float(line[29:].replace('D', 'E'))
-        # Gaussian uses the chemists notation for the 4-center indexes. HORTON
-        # uses the physicists notation.
-        set_four_index_element(result, i, k, j, l, value)
-    return result
+__all__ = ['FCHKFile', 'load_fchk']
 
 
 class FCHKFile(dict):
