+++ iodata/test/test_fchk.py
@@ -19,14 +19,14 @@
 #
 # --
 # pragma pylint: disable=invalid-name
-"""Test iodata.gaussian module."""
+"""Test iodata.fchk module."""
 
 
 import numpy as np
 
 from nose.tools import assert_raises
 
-from .. fchk import load_operators_g09, load_fchk
+from .. fchk import load_fchk
 from .. iodata import IOData
 from .. utils import shells_to_nbasis, check_dm
 from .. overlap import compute_overlap
@@ -35,80 +35,8 @@ try:
 except ImportError:
     from importlib.resources import path
 
+
 #TODO: shells_to_nbasis(obasis["shell_types"]) replacement test
-
-
-def test_load_operators_water_sto3g_hf_g03():
-    eps = 1e-5
-    with path('iodata.test.cached', 'water_sto3g_hf_g03.log') as fn:
-        result = load_operators_g09(str(fn))
-
-    overlap = result['olp']
-    kinetic = result['kin']
-    nuclear_attraction = result['na']
-    electronic_repulsion = result['er']
-
-    assert overlap.shape == (7, 7)
-    assert kinetic.shape == (7, 7)
-    assert nuclear_attraction.shape == (7, 7)
-    assert electronic_repulsion.shape == (7, 7, 7, 7)
-
-    assert abs(overlap[0, 0] - 1.0) < eps
-    assert abs(overlap[0, 1] - 0.236704) < eps
-    assert abs(overlap[0, 2] - 0.0) < eps
-    assert abs(overlap[-1, -3] - (-0.13198)) < eps
-
-    assert abs(kinetic[2, 0] - 0.0) < eps
-    assert abs(kinetic[4, 4] - 2.52873) < eps
-    assert abs(kinetic[-1, 5] - 0.00563279) < eps
-    assert abs(kinetic[-1, -3] - (-0.0966161)) < eps
-
-    assert abs(nuclear_attraction[3, 3] - 9.99259) < eps
-    assert abs(nuclear_attraction[-2, -1] - 1.55014) < eps
-    assert abs(nuclear_attraction[2, 6] - 2.76941) < eps
-    assert abs(nuclear_attraction[0, 3] - 0.0) < eps
-
-    assert abs(electronic_repulsion[0, 0, 0, 0] - 4.78506575204) < eps
-    assert abs(electronic_repulsion[6, 6, 6, 6] - 0.774605944194) < eps
-    assert abs(electronic_repulsion[6, 5, 0, 5] - 0.0289424337101) < eps
-    assert abs(electronic_repulsion[5, 4, 0, 1] - 0.00274145291476) < eps
-
-
-def test_load_operators_water_ccpvdz_pure_hf_g03():
-    eps = 1e-5
-    with path('iodata.test.cached', 'water_ccpvdz_pure_hf_g03.log') as fn:
-        result = load_operators_g09(str(fn))
-
-    overlap = result['olp']
-    kinetic = result['kin']
-    nuclear_attraction = result['na']
-    electronic_repulsion = result['er']
-
-    assert overlap.shape == (24, 24)
-    assert kinetic.shape == (24, 24)
-    assert nuclear_attraction.shape == (24, 24)
-    assert electronic_repulsion.shape == (24, 24, 24, 24)
-
-    assert abs(overlap[0, 0] - 1.0) < eps
-    assert abs(overlap[0, 1] - 0.214476) < eps
-    assert abs(overlap[0, 2] - 0.183817) < eps
-    assert abs(overlap[10, 16] - 0.380024) < eps
-    assert abs(overlap[-1, -3] - 0.000000) < eps
-
-    assert abs(kinetic[2, 0] - 0.160648) < eps
-    assert abs(kinetic[11, 11] - 4.14750) < eps
-    assert abs(kinetic[-1, 5] - (-0.0244025)) < eps
-    assert abs(kinetic[-1, -6] - (-0.0614899)) < eps
-
-    assert abs(nuclear_attraction[3, 3] - 12.8806) < eps
-    assert abs(nuclear_attraction[-2, -1] - 0.0533113) < eps
-    assert abs(nuclear_attraction[2, 6] - 0.173282) < eps
-    assert abs(nuclear_attraction[-1, 0] - 1.24131) < eps
-
-    assert abs(electronic_repulsion[0, 0, 0, 0] - 4.77005841522) < eps
-    assert abs(electronic_repulsion[23, 23, 23, 23] - 0.785718708997) < eps
-    assert abs(electronic_repulsion[23, 8, 23, 2] - (-0.0400337571969)) < eps
-    assert abs(electronic_repulsion[15, 2, 12, 0] - (-0.0000308196281033)) < eps
 
 
 def test_load_fchk_nonexistent():
